refactor(error_log): move ErroredEventContext::to_jsonl from pipeline.rs
